Nearly finished task 4, cleared a lot of unused files
Need to get the zumo to detect if it's turning back into a corridor it's already been to. Current code for this is buggy.

Need to comment the code in main.ino.

Removed unused libraries.

Added and commented the processing file which is used to communicate wirelessly to the arduino.